+++ packages/office-ui-fabric-react/src/components/ActivityItem/ActivityItem.classNames.ts
@@ -59,6 +59,9 @@ export const getClassNames = memoizeFunction((
 
     activityText: mergeStyles('ms-ActivityItem-activityText', styles.activityText),
     commentText: mergeStyles('ms-ActivityItem-commentText', styles.commentText),
-    timeStamp: mergeStyles('ms-ActivityItem-timeStamp', styles.timeStamp)
+    timeStamp: mergeStyles(
+      'ms-ActivityItem-timeStamp',
+      styles.timeStamp,
+      isCompact && styles.isCompactTimeStamp)
   };
 });