+++ packages/office-ui-fabric-react/src/components/Dialog/DialogFooter.types.ts
@@ -20,12 +20,9 @@ export interface IDialogFooterProps extends React.Props<DialogFooterBase> {
   theme?: ITheme;
 
   /**
-   * Additional css class to apply to the DialogFooter
-   * @defaultvalue undefined
-   */
+  * Optional override class name
+  */
   className?: string;
-
-  // Insert DialogFooter props below
 }
 
 export interface IDialogFooterStyleProps {
@@ -35,18 +32,17 @@ export interface IDialogFooterStyleProps {
   theme: ITheme;
 
   /**
-   * Accept custom classNames
-   */
+  * Optional override class name
+  */
   className?: string;
-
-  // Insert DialogFooter style props below
 }
 
 export interface IDialogFooterStyles {
   /**
-   * Style for the root element.
+   * Style for the actions element.
    */
-  actions?: IStyle;
+  actions: IStyle;
 
-  // Insert DialogFooter classNames below
+  actionsRight: IStyle;
+  action: IStyle;
 }