+++ packages/office-ui-fabric-react/src/components/SearchBox/SearchBox.base.tsx
@@ -2,11 +2,11 @@ import * as React from 'react';
 import { ISearchBoxProps, ISearchBoxStyleProps, ISearchBoxStyles } from './SearchBox.types';
 import {
   BaseComponent,
-  autobind,
   getId,
   KeyCodes,
   customizable,
-  classNamesFunction
+  classNamesFunction,
+  createRef
 } from '../../Utilities';
 
 import { IconButton } from '../../Button';
@@ -22,8 +22,8 @@ export interface ISearchBoxState {
 
 @customizable('SearchBox', ['theme'])
 export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxState> {
-  private _rootElement: HTMLElement;
-  private _inputElement: HTMLInputElement;
+  private _rootElement = createRef<HTMLDivElement>();
+  private _inputElement = createRef<HTMLInputElement>();
   private _latestValue: string;
 
   public constructor(props: ISearchBoxProps) {
@@ -77,7 +77,7 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
 
     return (
       <div
-        ref={ this._resolveRef('_rootElement') }
+        ref={ this._rootElement }
         className={ classNames.root }
         onFocusCapture={ this._onFocusCapture }
       >
@@ -94,7 +94,7 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
           value={ value }
           disabled={ this.props.disabled }
           aria-label={ ariaLabel ? ariaLabel : placeholder }
-          ref={ this._resolveRef('_inputElement') }
+          ref={ this._inputElement }
         />
         { value!.length > 0 &&
           <div className={ classNames.clearButton }>
@@ -114,9 +114,16 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
    * Sets focus to the search box input field
    */
   public focus() {
-    if (this._inputElement) {
-      this._inputElement.focus();
+    if (this._inputElement.value) {
+      this._inputElement.value.focus();
     }
+  }
+
+  /**
+   * Returns whether or not the SearchBox has focus
+   */
+  public hasFocus(): boolean {
+    return !!this.state.hasFocus;
   }
 
   private _onClear(ev: React.MouseEvent<HTMLAnchorElement | HTMLButtonElement> | React.KeyboardEvent<HTMLElement>) {
@@ -130,25 +137,23 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
       ev.stopPropagation();
       ev.preventDefault();
 
-      this._inputElement.focus();
+      this.focus();
     }
   }
 
-  @autobind
-  private _onFocusCapture(ev: React.FocusEvent<HTMLElement>) {
+  private _onFocusCapture = (ev: React.FocusEvent<HTMLElement>) => {
     this.setState({
       hasFocus: true
     });
 
-    this._events.on(this._rootElement, 'blur', this._onBlur, true);
+    this._events.on(this._rootElement.value, 'blur', this._onBlur, true);
 
     if (this.props.onFocus) {
       this.props.onFocus(ev as React.FocusEvent<HTMLInputElement>);
     }
   }
 
-  @autobind
-  private _onClearClick(ev: React.MouseEvent<HTMLAnchorElement | HTMLButtonElement>) {
+  private _onClearClick = (ev: React.MouseEvent<HTMLAnchorElement | HTMLButtonElement>) => {
     const { clearButtonProps } = this.props;
 
     if (clearButtonProps && clearButtonProps.onClick) {
@@ -160,8 +165,7 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
     }
   }
 
-  @autobind
-  private _onKeyDown(ev: React.KeyboardEvent<HTMLInputElement>) {
+  private _onKeyDown = (ev: React.KeyboardEvent<HTMLInputElement>) => {
 
     switch (ev.which) {
 
@@ -191,9 +195,8 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
     ev.stopPropagation();
   }
 
-  @autobind
-  private _onBlur(ev: React.FocusEvent<HTMLInputElement>) {
-    this._events.off(this._rootElement, 'blur');
+  private _onBlur = (ev: React.FocusEvent<HTMLInputElement>): void => {
+    this._events.off(this._rootElement.value, 'blur');
     this.setState({
       hasFocus: false
     });
@@ -203,8 +206,7 @@ export class SearchBoxBase extends BaseComponent<ISearchBoxProps, ISearchBoxStat
     }
   }
 
-  @autobind
-  private _onInputChange(ev: React.ChangeEvent<HTMLInputElement>) {
+  private _onInputChange = (ev: React.ChangeEvent<HTMLInputElement>) => {
     const value = ev.target.value;
 
     if (value === this._latestValue) {