+++ packages/file-type-icons/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/file-type-icons
 
-This log was last generated on Fri, 12 Jan 2018 20:03:22 GMT and should not be manually modified.
+This log was last generated on Tue, 13 Mar 2018 20:00:06 GMT and should not be manually modified.
+
+## 0.5.0
+Tue, 13 Mar 2018 20:00:06 GMT
+
+### Minor changes
+
+- Adjusting npmrc to ensure we have pu
+
+## 0.4.0
+Tue, 13 Mar 2018 10:17:37 GMT
+
+### Minor changes
+
+- Update file type icon mappings
 
 ## 0.3.0
 Fri, 12 Jan 2018 20:03:21 GMT